+++ src/enclave/tls_endpoint.h
@@ -19,7 +19,7 @@ namespace ccf
   {
   protected:
     ringbuffer::WriterPtr to_host;
-    int64_t session_id;
+    tls::ConnID session_id;
     size_t execution_thread;
 
     enum Status