+++ src/node/session_metrics.h
@@ -11,12 +11,18 @@ namespace ccf
 {
   struct SessionMetrics
   {
+    struct Errors
+    {
+      size_t parsing;
+    };
+
     struct PerInterface
     {
       size_t active;
       size_t peak;
       size_t soft_cap;
       size_t hard_cap;
+      Errors errors;
     };
 
     size_t active;
@@ -24,9 +30,12 @@ namespace ccf
     std::map<std::string, PerInterface> interfaces;
   };
 
+  DECLARE_JSON_TYPE(SessionMetrics::Errors)
+  DECLARE_JSON_REQUIRED_FIELDS(SessionMetrics::Errors, parsing)
+
   DECLARE_JSON_TYPE(SessionMetrics::PerInterface)
   DECLARE_JSON_REQUIRED_FIELDS(
-    SessionMetrics::PerInterface, active, peak, soft_cap, hard_cap)
+    SessionMetrics::PerInterface, active, peak, soft_cap, hard_cap, errors)
   DECLARE_JSON_TYPE(SessionMetrics)
   DECLARE_JSON_REQUIRED_FIELDS(SessionMetrics, active, peak, interfaces)
 }